code/plotly/active-prediction/src/requirements.txt
dash_html_components==0.13.2
dash==0.28.2
pandas==0.23.0
numpy==1.14.3
plotly==3.3.0
dash_table_experiments==0.6.0
dash_core_components==0.33.0
python_dateutil==2.7.3
dash-auth==1.2.0